Added support for episode memory, and searching by date
Added support to save the last episode of a show watched. Now if the episode number is not provided, it grabs from the list and plays the correct episode from the right season (not working by date right now). A .txt file will be created on first run

Added the option to specify "by date" or "from year" to find the correct content (ex. Suspiria 1977 vs Suspiria 2018).